fix id

+++ include/framework/entity.h
@@ -12,7 +12,7 @@
 
 class Entity {
   public:
-    explicit Entity(size_t _id = 0);
+    explicit Entity(uint16_t _id = 0);
     ~Entity();
 
     void add_component(IComponent* component);
@@ -28,10 +28,10 @@ class Entity {
             return nullptr;
         }
     }
-    size_t get_id();
-    void set_id(size_t _id);
+    uint16_t get_id();
+    void set_id(uint16_t _id);
   private:
-  	size_t id;
+  	uint16_t id;
   	std::list<INode*> nodes;
   	std::map<size_t, IComponent*> components;
 };
